requirements/requirements-optionals.txt
# Optional packages which may be used with REST framework.
psycopg2-binary>=2.8.5, <2.9
markdown
django-guardian
django-filter>=2.2.0, <2.3
coreapi
coreschema
pyyaml>=5.1
